pyproject.toml
[tool.poetry]
name = "ilm-red-api"
version = "1.0.0"
description = "Cloud-native API platform for digital knowledge management with AI-powered features"
authors = ["ILM Red Team"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.7"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pgvector = "^0.2.4"

# Validation & Settings
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}

# AI Providers (Multi-vendor support)
openai = "^1.10.0"
anthropic = "^0.14.0"
google-generativeai = "^0.4.0"  # Google Gemini
tiktoken = "^0.5.2"
# Note: Qwen (DashScope), xAI, and DeepSeek use httpx with OpenAI-compatible APIs

# Storage
azure-storage-blob = "^12.19.0"
aiofiles = "^23.2.1"

# PDF Processing
pypdf = "^3.17.0"
pdfplumber = "^0.10.3"
pymupdf = "^1.24.0"       # Page rendering (pure Python, no system deps)
pillow = "^10.2.0"        # Image processing for JPEG conversion

# Caching & Background Tasks
redis = "^5.0.1"
arq = "^0.25.0"

# Utilities
structlog = "^24.1.0"
httpx = "^0.26.0"
tenacity = "^8.2.3"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
factory-boy = "^3.3.0"
faker = "^22.0.0"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise exceptions with from (good practice, fix later)
    "UP007", # use X | Y for type annotations (cosmetic)
    "UP038", # use X | Y in isinstance (cosmetic)
    "E402",  # module level import not at top (sometimes intentional)
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
